Refactor VM and websocket modules for modularity
